requirements.txt
Django==2.2.8
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
